setup mac build; cleanup ms specific code; posix filesystem port; remove wide char support; exclude network and sound from build until everything else is ported; use glfw for window, gl context creation and user input; fix llvm specific warnings; use <chrono> for time; use fixed size types for serialization; a lot more...

+++ game/src/tank/gc/indicators.h
@@ -80,8 +80,8 @@ class GC_IndicatorBar : public GC_2dSprite
 protected:
 	ObjPtr<GC_2dSprite> _object;
 
-	DWORD _dwValue_offset;
-	DWORD _dwValueMax_offset;
+	size_t _dwValue_offset;
+	size_t _dwValueMax_offset;
 
 	LOCATION _location;
